% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{simulate_LP}
\alias{simulate_LP}
\title{Simulating with a Local Projection DGP}
\usage{
simulate_LP(
  M,
  T_,
  LP_lags,
  hmax,
  VAR_coefficients,
  Sigma_epsilon,
  irf_1to1,
  init_partial,
  z_quantiles,
  chi2_quantiles,
  selection,
  PIconstant,
  progress_bar,
  OLS,
  threads,
  seeds_gen,
  seeds_DL
)
}
\arguments{
\item{M}{integer number of simulation replications}

\item{T_}{integer sample size}

\item{LP_lags}{integer number of lags in the LP estimation}

\item{hmax}{integer maximum horizon}

\item{VAR_coefficients}{array/cube containing the VAR coefficients, with lags in the slices. These are output by the function make_VAR_coefs}

\item{Sigma_epsilon}{covariance matrix of the errors}

\item{irf_1to1}{vector vector containing the true impulse responses. These are output by the function irf_from_VAR}

\item{init_partial}{boolean whether the first step lasso should be only partly penalized}

\item{z_quantiles}{vector of quantiles from the normal distribution associated with the level of the desired test}

\item{chi2_quantiles}{vector of quantiles from the chi-squared distribution associated with the level of the desired test}

\item{selection}{integer in 1:4 which chooses the lambda selection method. 1: BIC, 2: AIC, 3: EBIC, 4: our suggested plug-in selector}

\item{PIconstant}{constant used in the plug-in estimator}

\item{progress_bar}{boolean whether a progress bar should be shown}

\item{OLS}{boolean whether OLS should be used for estimation instead of the desparsified lasso}

\item{threads}{integer number of cores used in parallel computation}

\item{seeds_gen}{vector of size M containing the random number seeds for generating the simulation DGP}

\item{seeds_DL}{array/cube of size 2x(hmax+1)xM containing the random number seeds for the plug-in lambda estimator of desparsified lasso}
}
\value{
Returns a list with the following elements: \cr
\item{\code{intervals}}{array/cube of confidence intervals of the impulse responses}
\item{\code{manual_Thetahat}}{array/cube of the Theta hat matrices}
\item{\code{betahats}}{array/cube of beta hats, the first stage lasso estimates}
}
\description{
Simulates a DGP based on the provided information and computes the impulse responses by local projections
}
